refactor: :truck: Split jwt models out into own files; claims, jwk & jwt
